// backend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  CUSTOMER: 'CUSTOMER',
  ORGANIZER: 'ORGANIZER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const TransactionStatus = {
  WAITING_PAYMENT: 'WAITING_PAYMENT',
  WAITING_CONFIRMATION: 'WAITING_CONFIRMATION',
  DONE: 'DONE',
  REJECTED: 'REJECTED',
  EXPIRED: 'EXPIRED',
  CANCELED: 'CANCELED'
} as const

export type TransactionStatus = (typeof TransactionStatus)[keyof typeof TransactionStatus]


export const DiscountType = {
  PERCENTAGE: 'PERCENTAGE',
  FIXED: 'FIXED'
} as const

export type DiscountType = (typeof DiscountType)[keyof typeof DiscountType]